# server/src/docs/recruiterPortal.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexus Partners — Recruiter Portal API",
...
    "version": "1.3.0",
    "contact": {
      "name": "Nexus Partners",
      "url": "https://nexuspartners.com"
    }
  },
  "servers": [
    {
      "url": "/api/recruiter",
      "description": "Recruiter API base path (prepend your server host, e.g. https://api.nexuspartners.com)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "RecruiterBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT from POST /auth/login. Expires per RECRUITER_JWT_EXPIRES_IN (default 12h)."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Invalid username or password" }
        },
        "required": ["error"]
      },
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "saved",
          "applied",
          "interview_scheduled",
          "interview_completed",
          "offer_received",
          "hired",
          "rejected",
          "dropped"
        ]
      },
...
      "StudentJobAction": {
        "type": "object",
        "nullable": true,
        "properties": {
          "status": { "$ref": "#/components/schemas/ApplicationStatus" },
          "statusLabel": { "type": "string", "example": "Applied" },
          "appliedAt": { "type": "string", "format": "date-time", "nullable": true },
          "droppedAt": { "type": "string", "format": "date-time", "nullable": true },
          "resumeFixedAt": { "type": "string", "format": "date-time", "nullable": true },
          "atsResumeUrl": { "type": "string", "format": "uri", "nullable": true }
        }
      },
...
      "ScrapedJob": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "theirstackJobId": { "type": "integer", "nullable": true },
          "source": { "type": "string", "enum": ["theirstack", "manual"] },
          "sourceLabel": {
            "type": "string",
            "example": "Internal Scraper",
            "description": "Display label for source"
          },
          "searchProfileId": { "type": "string", "nullable": true },
          "companyId": { "type": "string" },
          "jobTitle": { "type": "string", "example": "Senior Software Engineer" },
          "description": { "type": "string" },
          "companyName": { "type": "string", "example": "Acme Corp" },
          "companyDomain": { "type": "string" },
          "location": { "type": "string", "example": "Remote, US" },
          "countryCode": { "type": "string", "example": "US" },
          "remote": { "type": "boolean" },
          "hybrid": { "type": "boolean" },
          "workType": {
            "type": "string",
            "enum": ["Remote", "Hybrid", "Onsite"],
            "description": "Derived work type for job cards"
          },
          "seniority": { "type": "string" },
          "minExperienceYears": {
            "type": "number",
            "nullable": true,
            "description": "Minimum required experience in years (derived from seniority/text when not explicit)"
          },
          "maxExperienceYears": {
            "type": "number",
            "nullable": true,
            "description": "Maximum required experience in years (derived from seniority/text when not explicit)"
          },
          "technologySlugs": { "type": "array", "items": { "type": "string" } },
          "salaryMinUsd": { "type": "number", "nullable": true },
          "salaryMaxUsd": { "type": "number", "nullable": true },
          "salaryRange": {
            "type": "string",
            "example": "$120k–$150k",
            "description": "Formatted salary range; empty when unknown"
          },
          "postedAge": {
            "type": "string",
            "example": "3 days ago",
            "description": "Relative age from datePosted/createdAt"
          },
          "visaSponsorship": {
            "type": "boolean",
            "description": "Heuristic from job description/notes (H1B / sponsorship keywords). Same value as `isSponsored`."
          },
          "isSponsored": {
            "type": "boolean",
            "description": "Alias of `visaSponsorship`. Use for Sponsored badges/filters on job cards.",
            "example": true
          },
          "isApplied": {
            "type": "boolean",
            "description": "True when this student already applied (or progressed past apply) for this job. True for statuses: applied, interview_scheduled, interview_completed, offer_received, hired, rejected. False when there is no action or only saved.",
            "example": false
          },
          "tags": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["Remote", "H1B Sponsor"]
          },
          "metaLine": {
            "type": "string",
            "example": "Acme Corp · Remote · US · $120k–$150k · 3 days ago"
          },
          "datePosted": { "type": "string", "format": "date-time", "nullable": true },
          "discoveredAt": { "type": "string", "format": "date-time", "nullable": true },
          "applyUrl": { "type": "string", "format": "uri" },
          "finalUrl": { "type": "string", "format": "uri" },
          "sourceUrl": { "type": "string", "format": "uri" },
          "urlDomain": {
            "type": "string",
            "example": "boards.greenhouse.io",
            "description": "Scrape/board domain when available. Do not display on recruiter job cards; use `sourceLabel` only if a source badge is needed."
          },
          "notes": { "type": "string" },
          "status": { "type": "string", "enum": ["open", "applied", "closed", "archived"] },
          "lastSyncedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "verified": { "type": "boolean", "description": "Always true in recruiter responses" },
          "studentAction": { "$ref": "#/components/schemas/StudentJobAction" }
        }
      },
      "JobListResponse": {
        "type": "object",
        "properties": {
          "jobs": { "type": "array", "items": { "$ref": "#/components/schemas/ScrapedJob" } },
          "total": { "type": "integer", "example": 142 },
          "page": { "type": "integer", "example": 0 },
          "limit": { "type": "integer", "example": 20 }
        }
      },
...
    }
  },
  "paths": {
...
    "/jobs": {
      "get": {
        "tags": ["Jobs"],
        "summary": "List jobs for a student",
        "description": "Returns scraped open jobs matched to the student's **role** (title keywords from the assigned student's role, e.g. Software Engineer / Data Engineer) unless free-text `q` is provided. Excludes dropped jobs.\n\nEach job includes enriched card fields: `sourceLabel`, `salaryRange`, `postedAge`, `workType`, `visaSponsorship`, `isSponsored`, `isApplied`, `tags`, `metaLine`, plus `minExperienceYears` / `maxExperienceYears`.\n\n**UI guidance:** show `isApplied` / `isSponsored` badges on cards; do **not** display `urlDomain` (scrape source).\n\n**Experience filter:** `minExp` and/or `maxExp` (aliases `experienceMin` / `experienceMax`) keep jobs whose experience band overlaps the requested range.\n\n**Sponsored filter:** `sponsored` (alias `isSponsored`) as `true` or `false` filters by visa-sponsorship keywords in description/notes.",
        "operationId": "listJobs",
        "security": [{ "RecruiterBearerAuth": [] }],
        "parameters": [
          {
            "name": "studentPhone",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "example": "9876543210" }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index",
            "schema": { "type": "integer", "default": 0, "minimum": 0 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (max 50)",
            "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 50 }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Search title, company, location, description, or skills. When set, skips strict role matching.",
            "schema": { "type": "string" }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Filter by job source",
            "schema": { "type": "string", "enum": ["theirstack", "manual"] }
          },
          {
            "name": "remote",
            "in": "query",
            "description": "Filter remote jobs (`true` / `false`)",
            "schema": { "type": "string", "enum": ["true", "false"] }
          },
          {
            "name": "sponsored",
            "in": "query",
            "description": "Filter by visa sponsorship mention in job text. `true` = sponsored only; `false` = not sponsored. Alias: `isSponsored`.",
            "schema": { "type": "string", "enum": ["true", "false"], "example": "true" }
          },
          {
            "name": "isSponsored",
            "in": "query",
            "description": "Alias of `sponsored`",
            "schema": { "type": "string", "enum": ["true", "false"] }
          },
          {
            "name": "minExp",
            "in": "query",
            "description": "Minimum experience years filter. Returns jobs whose experience band overlaps this value (alias: experienceMin).",
            "schema": { "type": "number", "minimum": 0, "example": 2 }
          },
          {
            "name": "maxExp",
            "in": "query",
            "description": "Maximum experience years filter. Returns jobs whose experience band overlaps this value (alias: experienceMax).",
            "schema": { "type": "number", "minimum": 0, "example": 5 }
          },
          {
            "name": "experienceMin",
            "in": "query",
            "description": "Alias of `minExp`",
            "schema": { "type": "number", "minimum": 0 }
          },
          {
            "name": "experienceMax",
            "in": "query",
            "description": "Alias of `maxExp`",
            "schema": { "type": "number", "minimum": 0 }
          },
          {
            "name": "scrapedFrom",
            "in": "query",
            "description": "Filter jobs scraped on or after this date (YYYY-MM-DD)",
            "schema": { "type": "string", "format": "date", "example": "2026-07-01" }
          },
          {
            "name": "scrapedTo",
            "in": "query",
            "description": "Filter jobs scraped on or before this date (YYYY-MM-DD)",
            "schema": { "type": "string", "format": "date", "example": "2026-07-07" }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated job list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobListResponse" }
              }
            }
          },
          "400": { "description": "studentPhone required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Student not assigned", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}